LayoutTests/http/wpt/html/browsers/the-window-object/window-is-constructor-expected.txt


PASS Window is a constructor, and throwing TypeError when it is called. 
